requirements.txt
runpod>=1.7.0
Pillow>=10.0.0
requests>=2.31.0
huggingface_hub>=0.25.0
